TODO: separate join/create classroom modals into components
