+++ vue-frontend/src/views/UserHomeView.vue
@@ -58,12 +58,12 @@
       </div>
 
       <!-- Join Classroom (for classrooms no.>= 1)-->
-      <RouterLink v-if="classrooms.length && userRole === 'student'" :to="{ name: 'joinClassroom' }" class="hover:border-gray-100 hover:text-gray-100 ease-in-out duration-500 my-4 mx-4 rounded-md border-dashed border-2 border-gray-500 bg-transparent text-gray-500 h-48 sm:w-48 overflow-hidden w-ful">
+      <div v-if="classrooms.length && userRole === 'student'" :to="{ name: 'joinClassroom' }" class="hover:border-gray-100 hover:text-gray-100 ease-in-out duration-500 my-4 mx-4 rounded-md border-dashed border-2 border-gray-500 bg-transparent text-gray-500 h-48 sm:w-48 overflow-hidden w-ful">
         <div class="w-full h-full text-2xl flex flex-col justify-items-center justify-center">
           <p class="place-self-center font-bold">+</p>
           <p class="place-self-center font-bold text-center">Join More Classrooms</p>
         </div>
-      </RouterLink>
+      </div>
 
       <!-- Create Classroom -->
       <RouterLink v-else-if="classrooms.length && userRole === 'teacher'" :to="{ name: 'createClassroom' }" class="hover:border-gray-100 hover:text-gray-100 ease-in-out duration-500 my-4 mx-4 rounded-md border-dashed border-2 border-gray-500 bg-transparent text-gray-500 h-48 sm:w-48 overflow-hidden w-ful">
